Adding the Dual-use and offstore spyware apps to the blocklist.
