--- static_data/note.txt
@@ -1,2 +1,0 @@
-Add all these files to `static_data/` folder in your repo,
-and change `TEST=True` in `config.py`
